[ui] Rename Jobs page content
