+++ js_modules/dagster-ui/packages/ui-core/src/overview/OverviewRoot.tsx
@@ -1,13 +1,13 @@
 import {Redirect, Route, Switch} from 'react-router-dom';
 
 import {OverviewActivityRoot} from './OverviewActivityRoot';
-import {OverviewJobsRoot} from './OverviewJobsRoot';
 import {OverviewResourcesRoot} from './OverviewResourcesRoot';
 import {OverviewSchedulesRoot} from './OverviewSchedulesRoot';
 import {OverviewSensorsRoot} from './OverviewSensorsRoot';
 import {AutomaterializationRoot} from '../assets/auto-materialization/AutomaterializationRoot';
 import {InstanceBackfillsRoot} from '../instance/InstanceBackfillsRoot';
 import {BackfillPage} from '../instance/backfill/BackfillPage';
+import {OverviewJobsRoot} from '../jobs/JobsPageContent';
 
 export const OverviewRoot = () => {
   return (
